Back/requirements.txt
fastapi
uvicorn[standard]
firebase-admin
pydantic
python-dotenv
werkzeug
python-jose[cryptography]
python-multipart